pyproject.toml
[project]
name = "kurabye-bisquits"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]

[tool.ruff]
line-length = 120
src = ["app"]
# rooles: https://docs.astral.sh/ruff/rules/
select = [
	"E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "UP",   # pyupgrade
    "N",    # pep8-naming
	"UP",   # pyupgrade
	"N",    # pep8-naming
	"ERA",  # found commented-out code
	"D401", # check docstring format
]
unfixable = ["F401"]
ignore = [
 "B008",    # ruff swears at depends in the func
]
exclude = [".venv", ".git", "__pycache__", "build", "dist"]

[tool.black]
# Black is disabled by omitting its configuration here.

[tool.pyright]
# Delete if you dont have mistakes
executionEnvironments = [{ root = "app" }]
typeCheckingMode = "standard"
venv = "./.venv"
venvPath = "."
